# src/token/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "XRPL Token Insurance Demo API",
    "version": "0.4.0",
    "description": "REST API for XRPL Token Insurance Demo platform - stablecoin (JPYN) issuance with fiat↔token exchange"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Google OAuth2 ID Token"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" }
        }
      },
...
      "Token": {
        "type": "object",
        "properties": {
          "tokenId": { "type": "string" },
          "name": { "type": "string" },
          "currency": { "type": "string" },
          "domain": { "type": "string" },
          "issuerAddress": { "type": "string" },
          "createdAt": { "type": "string" }
        }
      },
...
      "ExchangeOrder": {
        "type": "object",
        "properties": {
          "orderId": { "type": "string" },
          "userId": { "type": "string" },
          "tokenId": { "type": "string" },
          "direction": { "type": "string", "enum": ["fiat_to_token", "token_to_fiat"] },
          "amount": { "type": "number" },
          "status": { "type": "string", "enum": ["pending", "fiat_debited", "token_burned", "completed", "failed"] },
          "xrplTxHash": { "type": "string" },
          "failureReason": { "type": "string" },
          "createdAt": { "type": "string" },
          "updatedAt": { "type": "string" }
        }
      },
...
      "PubSubEnvelope": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": {
            "type": "object",
            "required": ["data", "messageId"],
            "properties": {
              "data": { "type": "string", "description": "Base64-encoded JSON payload" },
              "messageId": { "type": "string" }
            }
          },
          "subscription": { "type": "string" }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/tokens": {
      "get": {
        "operationId": "listTokens",
        "summary": "List stablecoins",
        "tags": ["Token"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of tokens",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Token" }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/v1/tokens/{tokenId}": {
      "get": {
        "operationId": "getToken",
        "summary": "Get stablecoin details",
        "tags": ["Token"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "tokenId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Token details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Token" }
              }
            }
          },
          "404": {
            "description": "Token not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/exchange/fiat-to-xrp": {
      "post": {
        "operationId": "exchangeFiatToToken",
        "summary": "Exchange fiat to XRP token",
        "tags": ["Exchange"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tokenId", "fiatAmount"],
                "properties": {
                  "tokenId": { "type": "string" },
                  "fiatAmount": { "type": "number", "minimum": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Exchange order created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExchangeOrder" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/v1/exchange/xrp-to-fiat": {
      "post": {
        "operationId": "exchangeTokenToFiat",
        "summary": "Exchange XRP token to fiat",
        "tags": ["Exchange"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tokenId", "tokenAmount"],
                "properties": {
                  "tokenId": { "type": "string" },
                  "tokenAmount": { "type": "number", "minimum": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Exchange order created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExchangeOrder" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tokens/{tokenId}/trustline": {
      "post": {
        "operationId": "ensureTrustLine",
        "summary": "Ensure TrustLine exists for current user on a token",
        "tags": ["Token"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "tokenId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "TrustLine set (or already exists)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tokenId": { "type": "string" },
                    "currency": { "type": "string" },
                    "status": { "type": "string", "example": "ok" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Token or user not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/v1/pubsub/bank/deposit": {
      "post": {
        "operationId": "pubsubBankDeposit",
        "summary": "Pub/Sub push endpoint for bank deposit events",
        "tags": ["Pubsub"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PubSubEnvelope" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message acknowledged",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "reason": { "type": "string" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Processing failed (triggers Pub/Sub retry)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/v1/eventarc/xrpl/deposit": {
      "post": {
        "operationId": "eventarcXrplDeposit",
        "summary": "Eventarc Firestore trigger endpoint for XRPL token deposit events",
        "tags": ["Eventarc"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "document": {
                    "type": "object",
                    "properties": {
                      "name": { "type": "string" },
                      "fields": { "type": "object" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event processed or skipped",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "reason": { "type": "string" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Processing failed (triggers Eventarc retry)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    }
  }
}